refactor: split app utilities into modules
